Import all of the footprints into the layout
Delete the traces and footprints from the M.2 breakout board.

Update the design rules constraints and physical stackup using PCBWay's
manufacturing capabilities. These rules should work well for essentially
any manufacturer that ends up being used.

Adjust the NVT4857UK's WLCSP20 footprint according to NXP's recommended
footprint and PCBWay's manufacturing capabilities (allowing for at least
a 4mil soldermask web/bridge between pads). This package's small 0.4mm
pitch makes it so that via-in-pad technology is required for this
board's design.

Create a footprint which contains a MicroSD card 3D Step model and add
it to the board, specifying that it does not exist in the schematic.
This model is used to determine the placement of the MicroSD slot,
making sure that there is sufficient room to insert and remove the
MicroSD card while the M.2 module is installed in the phone. The current
position of x=131.8mm, y=-98.3mm for the DM3AT-SF-PEJM5 MicroSD slot allows
for this to be possible.