Finish the routing of the board and export the gerbers and all manufacturing files
Add fab notes to the layout.

Generate a fabrication drawing and include it in the gerbers directory
since that is the location PCBWay will use (when this directory's
contents are archived into a zip).

Export the 3D step model of the board.

Export the IPC-D-356 netlist file which will be used to verify the
gerbers.